feat: new stock reserve & consume method

+++ et/stock_erp/models/stock_moves_erp.py
@@ -22,42 +22,59 @@ class StockMovesERP(models.Model):
     picking_id = fields.Many2one('stock.picking')
     product_id = fields.Many2one('product.product')
     quantity = fields.Integer()
+    quantity_delivered = fields.Integer()
     bultos = fields.Float()
     uxb = fields.Integer()
+    type = fields.Selection(selection=[('reserve', 'Reserva'), ('delivery', 'Entrega')])
+
+    @api.model
+    def create(self, vals):
+        
+        records = super().create(vals)
+
+        for record in records:
+            if record.type == 'reserve':
+                record.stock_erp.increase_comprometido_unidades(record.quantity)
+            elif record.type == 'delivery':
+                record.stock_erp.decrease_fisico_unidades(record.quantity)
+                record.stock_erp.decrease_comprometido_unidades(record.quantity)
+
 
     def unreserve_stock(self):
         for record in self:
-            if not record.sale_id.picking_ids:
-                continue
+            if record.type == 'reserve':
+                if not record.sale_id.picking_ids:
+                    continue
 
-            partial_quantity = 0
-            total_qty_to_unlink = record.quantity
-            moves_to_unlink = []
-            pickings_to_check = set()
+                partial_quantity = 0
+                total_qty_to_unlink = record.quantity
+                moves_to_unlink = []
+                pickings_to_check = set()
 
-            for picking in record.sale_id.picking_ids:
-                if picking.state_wms != 'no':
-                    raise UserError(
-                        f'No se puede liberar la cantidad comprometida ({record.sale_id.name}) porque las unidades están siendo preparadas o ya fueron preparadas en la transferencia {picking.name}.'
-                    )
+                for picking in record.sale_id.picking_ids:
+                    if picking.state_wms != 'no':
+                        raise UserError(
+                            f'No se puede liberar la cantidad comprometida ({record.sale_id.name}) porque las unidades están siendo preparadas o ya fueron preparadas en la transferencia {picking.name}.'
+                        )
 
-                for move in picking.move_ids_without_package:
-                    if move.product_id.id != record.product_id.id:
-                        continue
+                    for move in picking.move_ids_without_package:
+                        if move.product_id.id != record.product_id.id:
+                            continue
 
-                    if move.product_uom_qty + partial_quantity <= total_qty_to_unlink:
-                        moves_to_unlink.append(move)
-                        partial_quantity += move.product_uom_qty
+                        if move.product_uom_qty + partial_quantity <= total_qty_to_unlink:
+                            moves_to_unlink.append(move)
+                            partial_quantity += move.product_uom_qty
 
-            if partial_quantity >= total_qty_to_unlink:
-                for move in moves_to_unlink:
-                    record.unreserve_sale_line(move)
-                    pickings_to_check.add(move.picking_id)
-                    move.state = 'draft'
-                    move.unlink()
-                    record.check_picking(pickings_to_check)
-                record.update_sale_orders()
-                record.unlink()
+                if partial_quantity >= total_qty_to_unlink:
+                    for move in moves_to_unlink:
+                        record.unreserve_sale_line(move)
+                        pickings_to_check.add(move.picking_id)
+                        move.state = 'draft'
+                        move.unlink()
+                        record.check_picking(pickings_to_check)
+                    record.stock_erp.decrease_comprometido_unidades(record.quantity)
+                    record.update_sale_orders()
+                    record.unlink()
 
     def update_sale_orders(self):
         for record in self:
